requirements.txt
Django>=2.0,<3.0
psycopg2==2.8.6
djangorestframework==3.12.4
django-filter==21.1